Use PG_TEST_TIMEOUT_DEFAULT in 019_replslot_limit.pl.
Oversight in 3ccdb08d1f5a5d792ccda11d1940c092c327a1fc, which introduced
the variable.  This lowers some 1000s timeouts to the configurable
default of 180s, due to a lack of evidence for needing the longer
timeout.  The alternative was 6*PG_TEST_TIMEOUT_DEFAULT, which we can
adopt if the need arises.  Given the lack of observed trouble with these
timeouts, no back-patch.