Revert "Add support for Kerberos credential delegation"
This reverts commit 787c19f248058d81c7d090db434fc7cd2029efdf.

Per discussion and buildfarm, this depends on APIs that seem to not
be available on at least one platform (NetBSD).  Should be certainly
possible to rework to be optional on that platform if necessary but bit
late for that at this point.

Discussion: https://postgr.es/m/3286097.1680922218@sss.pgh.pa.us